backend/requirements.txt
# Core FastAPI stack
fastapi
uvicorn
python-multipart

# Database
sqlalchemy

# Auth
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]

# LLM / AI
groq

# HTTP client
requests

# Utilities
python-dotenv
pydantic[email]

# Job scraping (LinkedIn, Naukri, Indeed, etc.)
python-jobspy

# MCP server
mcp[cli]